refactor: move createHeaderContent function from CatalogPage class
